feat: refactor agent and runtime modules, add crypto key management and MQTT compatibility
