+++ agent_runtime/__init__.py
@@ -1,0 +1,11 @@
+from agent_runtime.runtime import on_connect, on_message, start_workers, heartbeat_loop, set_public_key
+from agent_runtime.crypto import ensure_crypto_keys
+
+__all__ = [
+    "on_connect",
+    "on_message",
+    "start_workers",
+    "heartbeat_loop",
+    "set_public_key",
+    "ensure_crypto_keys",
+]
